requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud dependencies
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.13.0
google-cloud-discoveryengine==0.11.9
google-auth==2.25.2

# Document processing
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
beautifulsoup4==4.12.2
lxml==4.9.3

# ADK and AI
google-genai==0.2.0
langchain==0.1.0
langchain-google-vertexai==0.1.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
tenacity==8.2.3
tiktoken==0.5.2
